refactor: util.tsとコンポーネントから型定義を移動
- util.tsからGoogleUser型を削除
- Body, WelcomeMessage, DailyReportコンポーネントの型定義を移動
- 新しいtypesディレクトリからimportするよう変更
