+++ src/components/Body/index.tsx
@@ -1,18 +1,14 @@
 import React from 'react';
 import { Auth } from 'firebase/auth';
-import { GoogleUser, isValidEmail } from '../../util';
+import { isValidEmail } from '../../util';
+import { GoogleUser } from '../../types';
+import { BodyProps } from '../../types/components';
 import { config } from '../../config';
 import TimesEsa from '../TimesEsa';
 import SignInDialog from '../SignInDialog';
 import WelcomeMessage from '../WelcomeMessage';
 import SignOutButton from '../SignOutButton';
 
-export type BodyProps = {
-  hasUserLanded: boolean;
-  isSignedIn: boolean;
-  user: GoogleUser;
-  firebaseAuth: Auth | null;
-}
 
 export const Body: React.FC<BodyProps> = (props: BodyProps) => {
   const isShowSignedInDialog = (): boolean => {
